+++ scripts/prepare_eager_tsv.R
@@ -46,7 +46,7 @@ save_ind_tsv <- function(data, rename, output_dir, ...) {
  data %>% select(-individual.Full_Individual_Id) %>%  readr::write_tsv(file=paste0(ind_dir,"/",ind_id,".tsv")) ## Output structure can be changed here.
 
   ## Print Autorun_eager version to file
-  AE_version <- "1.1.1"
+  AE_version <- "1.1.2"
   cat(AE_version, file=paste0(ind_dir,"/autorun_eager_version.txt"), fill=T, append = F)
 }
 
@@ -124,7 +124,7 @@ complete_pandora_table <- join_pandora_tables(
   convert_all_ids_to_values(., con = con) %>%
   filter(sample.Ethically_culturally_sensitive == FALSE) ## Exclude ethically/culturally sensitive data. Conservative since it excludes NAs
 
-tibble_input_iids <- complete_pandora_table %>% filter(sequencing.Batch == sequencing_batch_id) %>% select(individual.Full_Individual_Id)
+tibble_input_iids <- complete_pandora_table %>% filter(sequencing.Run_Id == sequencing_batch_id) %>% select(individual.Full_Individual_Id)
 
 ## Pull information from pandora, keeping only matching IIDs and requested Sequencing types.
 results <- inner_join(complete_pandora_table, tibble_input_iids, by=c("individual.Full_Individual_Id"="individual.Full_Individual_Id")) %>%