+++ scripts/run_Eager.sh
@@ -1,10 +1,18 @@
 #!/usr/bin/env bash
 
+## Defaults
+rush=''
+array=''
+temp_file=''
+
 ## Flood execution. Useful for testing/fast processing of small batches.
 if [[ $1 == "-r" || $1 == "--rush" ]]; then
     rush="-bg"
-else
-    rush=''
+elif [[ $1 == '-a' || $1 == "--array" ]]; then
+    array='TRUE'
+    temp_file="/mnt/archgen/Autorun_eager/$(date +'%y%m%d_%H%M')_Autorun_eager_queue.txt"
+    ## Create new empty file with the correct naming, or flush contents of file if somehow it exists.
+    echo -n '' > ${temp_file}
 fi
 
 nxf_path="/home/srv_autoeager/conda/envs/autoeager/bin/"
@@ -53,6 +61,24 @@ for analysis_type in "SG" "TF"; do
         #### Always running with resume will ensure runs are only ever resumed instead of restarting.
         if [[ ${eager_input} -nt ${eager_output_dir}/multiqc/multiqc_report.html ]]; then
 
+            if [[ ${array} == 'TRUE' ]]; then
+            ## For array submissions, the commands to be run will be added one by one to the temp_file
+            ## Then once all jobs have been added, submit that to qsub with each line being its own job.
+            ## Use `continue` to avoid running eager interactivetly for arrayed jobs.
+                echo "cd $(dirname ${eager_input}) ; ${nxf_path}/nextflow run nf-core/eager \
+                -r ${eager_version} \
+                -profile ${analysis_profiles} \
+                -c ${autorun_config} \
+                --input ${eager_input} \
+                --outdir ${eager_output_dir} \
+                -w ${eager_output_dir}/work \
+                -with-tower \
+                -ansi-log false \
+                ${run_name} ${rush}" | tr -s " " >> ${temp_file}
+                continue ## Skip running eager interactively if arrays are requested.
+            fi
+
+            ## NON-ARRAY RUNS
             ## Change to input directory to run from, to keep one cwd per run.
             cd $(dirname ${eager_input})
             ## Debugging info.
@@ -68,6 +94,7 @@ for analysis_type in "SG" "TF"; do
                 -ansi-log false \
                 ${run_name} ${rush}"
             
+
             ## Actually run eager now.
                 ## Monitor run in nf tower. Only works if TOWER_ACCESS_TOKEN is set.
                 ## Runs show in the Autorun_Eager workspace on tower.nf
@@ -86,3 +113,24 @@ for analysis_type in "SG" "TF"; do
         fi
     done
 done
+
+## If array is requested submit the created array file to qsub below
+if [[ ${array} == 'TRUE' ]]; then
+    mkdir -p array_Logs/$(basename ${temp_file}) ## Create new directory for the logs for more traversable structure
+    jn=$(wc -l ${temp_file} | cut -f 1 -d " ") ## number of jobs equals number of lines
+    export NXF_OPTS='-Xms4G -Xmx4G' ## Set 4GB limit to Nextflow VM
+    export JAVA_OPTS='-Xms8G -Xmx8G' ## Set 8GB limit to Java VM
+    ## -V Pass environment to job (includes nxf/java opts)
+    ## -S /bin/bash Use bash
+    ## -l v_hmem=30G ## 24GB memory limit (8 for java + the rest for garbage collector)
+    ## -pe smp 2 ## Use two cores. one for nextflow, one for garbage collector
+    ## -n AE_spawner ## Name the job
+    ## -cwd Run in currect run directory (ran commands include a cd anyway, but to find the files at least)
+    ## -j y ## join stderr and stdout into one output log file
+    ## -b y ## Provided command is a binary already (i.e. executable)
+    ## -o /mnt/archgen/Autorun_eager/array_Logs/ ## Keep all log files in one directory.
+    ## -tc 10 ## Number of concurrent spawner jobs (10)
+    ## -t 1-${jn} ## The number of array jobs (from 1 to $jn)
+    echo "qsub -V -S /bin/bash -l h_vmem=24G -pe smp 2 -N AE_spawner_$(basename ${temp_file}) -cwd -j y -b y -o /mnt/archgen/Autorun_eager/array_Logs/$(basename ${temp_file}) -tc 10 -t 1-${jn} /mnt/archgen/Autorun_eager/scripts/submit_as_array.sh ${temp_file}"
+    qsub -V -S /bin/bash -l h_vmem=30G -pe smp 2 -N AE_spawner_$(basename ${temp_file}) -cwd -j y -b y -o /mnt/archgen/Autorun_eager/array_Logs/$(basename ${temp_file}) -tc 10 -t 1-${jn} /mnt/archgen/Autorun_eager/scripts/submit_as_array.sh ${temp_file}
+fi