+++ CHANGELOG.md
@@ -3,6 +3,26 @@
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.1.2] - 02/01/2023
+
+### `Added`
+- Add `-a` option to `run_Eager.sh` to create a text file with the run commands for launching Autorun_eager, for submission as a qsub array.
+- Add array submission script.
+- Add script to create and update poseidon packages from eager output.
+  - Includes script to collate eager results and overwrite the janno of a poseidon pacakge.
+  - Includes script to update poseidon indFile from janno files.
+  - Each poseidon package contains a file named `AE_version.txt` with the version used for the last package creation/update.
+- Added conda environment yaml file `autoeager_env.yml`
+
+### `Fixed`
+- Fixed pull request template.
+- Fixed a bug where sequencing runs were considered updated if the `Results.txt` was updated without changing the data. Now only changes to the output bams constitute an update trigger.
+- Fixed a bug where TSV creation would try to match the sequencing ID to the sequencing batch ID, instead of the run ID.
+
+### `Dependencies`
+
+### `Deprecated`
+
 ## [1.1.1] - 11/10/2022
 
 ### `Added`